+++ src/strojure/ring_undertow/server.clj
@@ -1,21 +1,9 @@
 (ns strojure.ring-undertow.server
   (:require [strojure.ring-undertow.handler :as ring.handler]
             [strojure.undertow.api.types :as types]
-            [strojure.undertow.server :as server])
-  (:import (clojure.lang MultiFn)))
+            [strojure.undertow.server :as server]))
 
 (set! *warn-on-reflection* true)
-
-;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
-
-(defmulti handler-type-adapter
-  "Returns handler function adapter for `:ring-handler-type` key in server
-  configuration."
-  :ring-handler-type)
-
-(.addMethod ^MultiFn handler-type-adapter nil (constantly ring.handler/sync))
-(.addMethod ^MultiFn handler-type-adapter :sync (constantly ring.handler/sync))
-(.addMethod ^MultiFn handler-type-adapter :async (constantly ring.handler/async))
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
@@ -25,8 +13,9 @@
   Additional configurations keys.
 
   - `:ring-handler-type` The type of all ring handlers in the configuration.
-      - No value or value `:sync` means synchronous ring handler.
-      - Value `:async` means asynchronous ring handler.
+      - No value or value `:sync` means synchronous ring handler
+        [[handler/sync]].
+      - Value `:async` means asynchronous ring handler [[handler/async]].
 
   To use handlers of mixed type you can either:
 
@@ -36,7 +25,7 @@
   [1]: https://cljdoc.org/d/com.github.strojure/undertow/CURRENT/api/strojure.undertow.server#start
   "
   [config]
-  (binding [types/*handler-fn-adapter* (handler-type-adapter config)]
+  (binding [types/*handler-fn-adapter* (ring.handler/adapter config)]
     (server/start config)))
 
 (def ^{:arglists '([instance])} stop